fix: use mapper in raincloud
